Begin converting chart-controller to TypeScript
