+++ frontend/src/scripts/test/result.ts
@@ -13,7 +13,6 @@ import * as GlarsesMode from "../states/glarses-mode";
 import * as TestInput from "./test-input";
 import * as Notifications from "../elements/notifications";
 import { Chart } from "chart.js";
-import { AnnotationOptions } from "chartjs-plugin-annotation";
 import { Auth } from "../firebase";
 
 let result: MonkeyTypes.Result<MonkeyTypes.Mode>;
@@ -27,7 +26,7 @@ export function toggleUnsmoothedRaw(): void {
 }
 
 async function updateGraph(): Promise<void> {
-  ChartController.result.options.plugins!.annotation!.annotations = [];
+  ChartController.result.options.plugins.annotation.annotations = [];
   const labels = [];
   for (let i = 1; i <= TestInput.wpmHistory.length; i++) {
     if (TestStats.lastSecondNotRound && i === TestInput.wpmHistory.length) {
@@ -36,10 +35,9 @@ async function updateGraph(): Promise<void> {
       labels.push(i.toString());
     }
   }
-  (ChartController.result.data.labels as string[]) = labels;
-  (ChartController.result as Chart<"line" | "scatter">).options.scales![
-    "wpm"
-  ]!.title!.text = Config.alwaysShowCPM
+  ChartController.result.updateColors();
+  ChartController.result.data.labels = labels;
+  ChartController.result.options.scales["wpm"].title.text = Config.alwaysShowCPM
     ? "Character per Minute"
     : "Words per Minute";
   const chartData1 = Config.alwaysShowCPM
@@ -61,8 +59,8 @@ async function updateGraph(): Promise<void> {
       : result.chartData.raw;
   }
 
-  (ChartController.result.data.datasets[0].data as number[]) = chartData1;
-  (ChartController.result.data.datasets[1].data as number[]) = chartData2;
+  ChartController.result.data.datasets[0].data = chartData1;
+  ChartController.result.data.datasets[1].data = chartData2;
 
   ChartController.result.data.datasets[0].label = Config.alwaysShowCPM
     ? "cpm"
@@ -73,15 +71,14 @@ async function updateGraph(): Promise<void> {
     const minChartVal = Math.min(
       ...[Math.min(...chartData2), Math.min(...chartData1)]
     );
-    ChartController.result.options.scales!["wpm"]!.min = minChartVal;
-    ChartController.result.options.scales!["raw"]!.min = minChartVal;
+    ChartController.result.options.scales["wpm"].min = minChartVal;
+    ChartController.result.options.scales["raw"].min = minChartVal;
   } else {
-    ChartController.result.options.scales!["wpm"]!.min = 0;
-    ChartController.result.options.scales!["raw"]!.min = 0;
+    ChartController.result.options.scales["wpm"].min = 0;
+    ChartController.result.options.scales["raw"].min = 0;
   }
 
-  (ChartController.result.data.datasets[2].data as number[]) =
-    result.chartData.err;
+  ChartController.result.data.datasets[2].data = result.chartData.err;
 
   const fc = await ThemeColors.get("sub");
   if (Config.funbox !== "none") {
@@ -89,15 +86,12 @@ async function updateGraph(): Promise<void> {
     if (Config.funbox === "layoutfluid") {
       content += " " + Config.customLayoutfluid.replace(/#/g, " ");
     }
-    (
-      ChartController.result.options.plugins!.annotation!
-        .annotations as AnnotationOptions[]
-    ).push({
+    ChartController.result.options.plugins.annotation.annotations.push({
       display: true,
       id: "funbox-label",
       type: "line",
       scaleID: "wpm",
-      value: ChartController.result.options.scales!["wpm"]!.min,
+      value: ChartController.result.options.scales["wpm"].min,
       borderColor: "transparent",
       borderWidth: 1,
       borderDash: [2, 2],
@@ -120,10 +114,11 @@ async function updateGraph(): Promise<void> {
     });
   }
 
-  ChartController.result.options.scales!["wpm"]!.max = maxChartVal;
-  ChartController.result.options.scales!["raw"]!.max = maxChartVal;
-  ChartController.result.options.scales!["error"]!.max =
-    Math.max(...result.chartData.err) + 1;
+  ChartController.result.options.scales["wpm"].max = maxChartVal;
+  ChartController.result.options.scales["raw"].max = maxChartVal;
+
+  ChartController.result.updateColors();
+  ChartController.result.resize();
 }
 
 export async function updateGraphPBLine(): Promise<void> {
@@ -141,11 +136,7 @@ export async function updateGraphPBLine(): Promise<void> {
   const chartlpb = Misc.roundTo2(Config.alwaysShowCPM ? lpb * 5 : lpb).toFixed(
     2
   );
-
-  (
-    ChartController.result.options.plugins!.annotation!
-      .annotations as AnnotationOptions[]
-  ).push({
+  ChartController.result.options.plugins.annotation.annotations.push({
     display: true,
     type: "line",
     id: "lpb",
@@ -177,12 +168,9 @@ export async function updateGraphPBLine(): Promise<void> {
   ) {
     maxChartVal = parseFloat(chartlpb) + 15;
   }
-  ChartController.result.options.scales!["wpm"]!.max = Math.round(
-    maxChartVal + 5
-  );
-  ChartController.result.options.scales!["raw"]!.max = Math.round(
-    maxChartVal + 5
-  );
+  ChartController.result.options.scales["wpm"].max = Math.round(maxChartVal);
+  ChartController.result.options.scales["raw"].max = Math.round(maxChartVal);
+  ChartController.result.updateColors();
 }
 
 function updateWpmAndAcc(): void {
@@ -424,10 +412,7 @@ function updateTags(dontSave: boolean): void {
         // console.log("new pb for tag " + tag.name);
       } else {
         const themecolors = await ThemeColors.getAll();
-        (
-          ChartController.result.options.plugins!.annotation!
-            .annotations as AnnotationOptions[]
-        ).push({
+        ChartController.result.options.plugins.annotation.annotations.push({
           display: true,
           type: "line",
           id: "tpb",
@@ -442,8 +427,8 @@ function updateTags(dontSave: boolean): void {
               family: Config.fontFamily.replace(/_/g, " "),
               size: 11,
               style: "normal",
-              weight: Chart.defaults.font.weight!,
-              lineHeight: Chart.defaults.font.lineHeight!,
+              weight: Chart.defaults.font.weight,
+              lineHeight: Chart.defaults.font.lineHeight,
             },
             color: themecolors["bg"],
             padding: 3,
